requirements.txt
beautifulsoup4==4.6.0
botocore==1.7.43
bs4==0.0.1
cachetools==2.0.1
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
configparser==3.5.0
cycler==0.10.0
dill==0.2.7.1
Django==1.11.3
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum==0.4.6
enum34==1.1.6
inflection==0.3.1
jmespath==0.9.3
jsonschema==2.5.1
lxml==4.1.1
pyparsing==2.2.0
pyreadline==2.1
python-dateutil==2.6.1
python-http-client==2.4.0
requests==2.18.4
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
